+++ test/roda/list_grid_config_test.rb
@@ -78,6 +78,12 @@ class ListGridConfigTest < Minitest::Test
     assert_equal :standard, config.conditions_key
     assert_equal 1, config.conditions.length
     assert_equal 'id', config.conditions.first[:col]
+
+    # Test when params are not part of querystring:
+    config = Crossbeams::DataGrid::ListGridConfig.new(id: 'agrid', root_path: '/a/b/c', params: { key: 'standard', sub_type_id: '3', product_column_ids: '[73, 74]' }, config_loader: loader_extended(additions))
+    assert_equal :standard, config.conditions_key
+    assert_equal 1, config.conditions.length
+    assert_equal 'id', config.conditions.first[:col]
   end
 
   def test_multisel_conditions