+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Check all SQL snippets (get caption, preselect ids, hide button etc.) are SELECTs and raise an exception if they are not.
 ### Fixed
+- Multiselect conditions code had a couple of problems.
 
 ## [0.1.8] - 2018-08-31
 ### Added